refactor: ディレクトリパス修正
