+++ .storybook/preview.tsx
@@ -2,7 +2,7 @@ import type { Preview } from "@storybook/react";
 import "../src/app/globals.css";
 // @ts-ignore
 // TS2307: Cannot find module '@/components/ui/toaster' or its corresponding type declarations.
-import { Toaster } from "@/__shared__/ui/shadcn/toaster";
+import { Toaster } from "@/__shared__/components/ui/toaster";
 import { initialize, mswLoader } from "msw-storybook-addon";
 
 // Initialize MSW
