+++ package.json
@@ -26,6 +26,7 @@
     "@hookform/resolvers": "^3.1.0",
     "@next-auth/prisma-adapter": "^1.0.7",
     "@prisma/client": "^4.15.0",
+    "@radix-ui/react-checkbox": "^1.0.4",
     "@radix-ui/react-dialog": "^1.0.4",
     "@radix-ui/react-radio-group": "^1.1.3",
     "@radix-ui/react-scroll-area": "^1.0.4",
@@ -38,7 +39,7 @@
     "@types/react-dom": "18.2.4",
     "autoprefixer": "10.4.14",
     "browser-image-compression": "^2.0.2",
-    "class-variance-authority": "^0.6.0",
+    "class-variance-authority": "^0.6.1",
     "clsx": "^1.2.1",
     "eslint": "8.41.0",
     "eslint-config-next": "13.4.4",
@@ -50,12 +51,13 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.44.2",
+    "recharts": "^2.7.2",
     "request": "^2.88.2",
     "retry-request": "^5.0.2",
     "sharp": "^0.32.1",
-    "tailwind-merge": "^1.12.0",
+    "tailwind-merge": "^1.13.2",
     "tailwindcss": "3.3.2",
-    "tailwindcss-animate": "^1.0.5",
+    "tailwindcss-animate": "^1.0.6",
     "typescript": "5.0.4",
     "ulid": "^2.3.0",
     "zod": "^3.21.4"
