+++ src/app/(operation-check)/operation-check/address/_components/address-form/address-select-dialog/index.tsx
@@ -5,9 +5,12 @@ import {
   DialogDescription,
   DialogHeader,
   DialogTitle,
-} from "@/__shared__/ui/shadcn/dialog";
-import { Label } from "@/__shared__/ui/shadcn/label";
-import { RadioGroup, RadioGroupItem } from "@/__shared__/ui/shadcn/radio-group";
+} from "@/__shared__/components/ui/dialog";
+import { Label } from "@/__shared__/components/ui/label";
+import {
+  RadioGroup,
+  RadioGroupItem,
+} from "@/__shared__/components/ui/radio-group";
 import { useAtom } from "jotai";
 import { isAddressSelectDialogOpenAtom, selectedAddressAtom } from "../";
 import { IAddress } from "../address-interface";
