+++ src/app/layout.tsx
@@ -1,5 +1,7 @@
-import { Toaster } from "@/__shared__/ui/shadcn/toaster";
+import { Toaster } from "@/__shared__/components/ui/toaster";
+import { url } from "@/__shared__/utils/url";
 import { Inter } from "next/font/google";
+import Link from "next/link";
 import React from "react";
 import "./globals.css";
 
@@ -19,7 +21,11 @@ export default function RootLayout({
     <html lang="ja">
       <body className={inter.className}>
         <div className="relative flex min-h-screen flex-col">
-          <div className={"flex h-16 items-center bg-gray-100"}>header</div>
+          <div className={"flex h-16 items-center bg-gray-100"}>
+            <Link href={url.TOP}>
+              <h1 className={"pl-8 font-extrabold"}>Next-Template</h1>
+            </Link>
+          </div>
           <main className="flex-1 space-y-8 p-24">{children}</main>
         </div>
